replace double by float (closes #13) (#100)
* closes #16

* also make remotes use longs

* several fixes; cf. PR comments

* yet another fix

* more polishing

* polishing

* polishing

* polishing

* all set; including test refinement

* fixing identation

* fix

* protecting against NPE

* fix

+++ broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/services/remote/dataobjects/RestBankOperationData.java
@@ -3,12 +3,13 @@ package pt.ulisboa.tecnico.softeng.broker.services.remote.dataobjects;
 import org.joda.time.DateTime;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import pt.ulisboa.tecnico.softeng.broker.domain.Adventure;
 
 public class RestBankOperationData {
 	private String reference;
 	private String type;
 	private String iban;
-	private Double value;
+	private long value;
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss.SSS")
 	private DateTime time;
 	private String transactionSource;
@@ -17,7 +18,7 @@ public class RestBankOperationData {
 	public RestBankOperationData() {
 	}
 
-	public RestBankOperationData(String iban, double value, String transactionSource, String transactionReference) {
+	public RestBankOperationData(String iban, long value, String transactionSource, String transactionReference) {
 		this.iban = iban;
 		this.value = value;
 		this.transactionSource = transactionSource;
@@ -48,11 +49,11 @@ public class RestBankOperationData {
 		this.iban = iban;
 	}
 
-	public Double getValue() {
+	public long getValue() {
 		return this.value;
 	}
 
-	public void setValue(Double value) {
+	public void setValue(long value) {
 		this.value = value;
 	}
