+++ broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/services/remote/dataobjects/RestRoomBookingData.java
@@ -18,7 +18,7 @@ public class RestRoomBookingData {
     private LocalDate departure;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     private LocalDate cancellationDate;
-    private double price;
+    private long price;
     private String paymentReference;
     private String invoiceReference;
     private String buyerNif;
@@ -110,11 +110,11 @@ public class RestRoomBookingData {
         this.cancellationDate = cancellationDate;
     }
 
-    public double getPrice() {
+    public long getPrice() {
         return this.price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(long price) {
         this.price = price;
     }
